SHA2-256 and SHA2-512 components (together with related sub-components) moved from sha256 to sha2 directory. Shared functionality of SHA2-256 and SHA2-512 underlying block processing components moved to a detail file named split_functions.hpp
