+++ include/nil/blueprint/components/hashes/sha2/plonk/sha256_process.hpp
@@ -35,6 +35,7 @@
 #include <nil/blueprint/blueprint/plonk/circuit.hpp>
 #include <nil/blueprint/blueprint/plonk/assignment.hpp>
 #include <nil/blueprint/component.hpp>
+#include <nil/blueprint/components/hashes/sha2/plonk/detail/split_functions.hpp>
 
 namespace nil {
     namespace blueprint {
@@ -581,138 +582,6 @@ namespace nil {
                     bp.add_gate(first_selector_index + 6,
                                 {constraint_out_1, constraint_out_2, constraint_out_3, constraint_out_4});
                 }
-
-                template<typename BlueprintFieldType>
-                std::array<std::vector<typename BlueprintFieldType::integral_type>, 2>
-                    split_and_sparse(std::vector<bool> bits, const std::vector<size_t> &sizes, std::size_t base) {
-
-                    std::size_t size = sizes.size() - 1;
-                    std::array<std::vector<typename BlueprintFieldType::integral_type>, 2> res = {
-                        std::vector<typename BlueprintFieldType::integral_type>(size + 1),
-                        std::vector<typename BlueprintFieldType::integral_type>(size + 1)};
-                    std::size_t k = 0;
-                    for (int i = size; i > -1; i--) {
-                        res[0][i] = int(bits[k]);
-                        res[1][i] = int(bits[k]);
-                        for (std::size_t j = 1; j < sizes[i]; j++) {
-                            res[0][i] = res[0][i] * 2 + int(bits[k + j]);
-                            res[1][i] = res[1][i] * base + int(bits[k + j]);
-                        }
-                        k = k + sizes[i];
-                    }
-                    return res;
-                }
-
-                template<typename BlueprintFieldType>
-                std::array<std::vector<typename BlueprintFieldType::integral_type>, 2>
-                    reversed_sparse_and_split(typename BlueprintFieldType::integral_type sparse_value,
-                                              const std::vector<size_t> &sizes, std::size_t base) {
-                    std::size_t size = sizes.size();
-                    std::array<std::vector<typename BlueprintFieldType::integral_type>, 2> res = {
-                        std::vector<typename BlueprintFieldType::integral_type>(size),
-                        std::vector<typename BlueprintFieldType::integral_type>(size)};
-                    typename BlueprintFieldType::integral_type sparse_base = base;
-                    typename BlueprintFieldType::value_type value_base = base;
-                    std::size_t k = -1;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        k = k + sizes[i];
-                    }
-                    typename BlueprintFieldType::integral_type tmp = sparse_value;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        res[0][i] = 0;
-                        res[1][i] = 0;
-                        for (int j = sizes[i] - 1; j > -1; j--) {
-                            if (tmp > typename BlueprintFieldType::integral_type(value_base.pow(k).data) - 1) {
-                                typename BlueprintFieldType::integral_type r =
-                                    (tmp - (tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data))) /
-                                    typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                                res[0][i] = res[0][i] * 2 + (r & 1);
-                                res[1][i] = res[1][i] * sparse_base + r;
-                            } else {
-                                res[0][i] = res[0][i] * 2;
-                                res[1][i] = res[1][i] * sparse_base;
-                            }
-                            tmp = tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                            k--;
-                        }
-                    }
-                    return res;
-                }
-
-                template<typename BlueprintFieldType>
-                std::array<std::vector<typename BlueprintFieldType::integral_type>, 2>
-                    reversed_sparse_and_split_maj(typename BlueprintFieldType::integral_type sparse_value,
-                                                  const std::vector<size_t> &sizes, std::size_t base) {
-                    std::size_t size = sizes.size();
-                    std::array<std::vector<typename BlueprintFieldType::integral_type>, 2> res = {
-                        std::vector<typename BlueprintFieldType::integral_type>(size),
-                        std::vector<typename BlueprintFieldType::integral_type>(size)};
-                    typename BlueprintFieldType::integral_type sparse_base = base;
-                    typename BlueprintFieldType::value_type value_base = base;
-                    std::size_t k = -1;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        k = k + sizes[i];
-                    }
-                    std::array<std::size_t, 4> r_values = {0, 0, 1, 1};
-                    typename BlueprintFieldType::integral_type tmp = sparse_value;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        res[0][i] = 0;
-                        res[1][i] = 0;
-                        for (int j = sizes[i] - 1; j > -1; j--) {
-                            if (tmp > typename BlueprintFieldType::integral_type(value_base.pow(k).data) - 1) {
-                                typename BlueprintFieldType::integral_type r =
-                                    (tmp - (tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data))) /
-                                    typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                                res[0][i] = res[0][i] * 2 + r_values[std::size_t(r)];
-                                res[1][i] = res[1][i] * sparse_base + r;
-                            } else {
-                                res[0][i] = res[0][i] * 2;
-                                res[1][i] = res[1][i] * sparse_base;
-                            }
-                            tmp = tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                            k--;
-                        }
-                    }
-                    return res;
-                }
-
-                template<typename BlueprintFieldType>
-                std::array<std::vector<typename BlueprintFieldType::integral_type>, 2>
-                    reversed_sparse_and_split_ch(typename BlueprintFieldType::integral_type sparse_value,
-                                                 const std::vector<size_t> &sizes, std::size_t base) {
-                    std::size_t size = sizes.size();
-                    std::array<std::vector<typename BlueprintFieldType::integral_type>, 2> res = {
-                        std::vector<typename BlueprintFieldType::integral_type>(size),
-                        std::vector<typename BlueprintFieldType::integral_type>(size)};
-                    typename BlueprintFieldType::integral_type sparse_base = base;
-                    typename BlueprintFieldType::value_type value_base = base;
-                    std::size_t k = -1;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        k = k + sizes[i];
-                    }
-                    std::array<std::size_t, 6> r_values = {0, 0, 1, 0, 1, 1};
-                    typename BlueprintFieldType::integral_type tmp = sparse_value;
-                    for (int i = sizes.size() - 1; i > -1; i--) {
-                        res[0][i] = 0;
-                        res[1][i] = 0;
-                        for (int j = sizes[i] - 1; j > -1; j--) {
-                            if (tmp > typename BlueprintFieldType::integral_type(value_base.pow(k).data) - 1) {
-                                typename BlueprintFieldType::integral_type r =
-                                    (tmp - (tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data))) /
-                                    typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                                res[0][i] = res[0][i] * 2 + r_values[std::size_t(r) - 1];
-                                res[1][i] = res[1][i] * sparse_base + r;
-                            } else {
-                                res[0][i] = res[0][i] * 2;
-                                res[1][i] = res[1][i] * sparse_base;
-                            }
-                            tmp = tmp % typename BlueprintFieldType::integral_type(value_base.pow(k).data);
-                            k--;
-                        }
-                    }
-                    return res;
-                }
-
             }    // namespace detail
 
             template<typename BlueprintFieldType, typename ArithmetizationParams>
@@ -1253,4 +1122,4 @@ namespace nil {
     }        // namespace blueprint
 }    // namespace nil
 
-#endif    // CRYPTO3_BLUEPRINT_COMPONENTS_PLONK_SHA256_HPP
+#endif    // CRYPTO3_BLUEPRINT_COMPONENTS_PLONK_SHA256_PROCESS_HPP
