+++ include/nil/blueprint/components/hashes/sha2/r1cs/sha256_construction.hpp
@@ -32,7 +32,7 @@
 
 #include <nil/blueprint/components/packing.hpp>
 #include <nil/blueprint/components/hashes/hash_io.hpp>
-#include <nil/blueprint/components/hashes/sha256/sha256_aux.hpp>
+#include <nil/blueprint/components/hashes/sha256/r1cs/sha256_aux.hpp>
 #include <nil/crypto3/zk/blueprint/r1cs.hpp>
 
 namespace nil {
