+++ include/nil/blueprint/components/merkle_tree/plonk/merkle_tree.hpp
@@ -31,7 +31,7 @@
 
 #include <nil/blueprint/blueprint/plonk/circuit.hpp>
 #include <nil/blueprint/blueprint/plonk/assignment.hpp>
-#include <nil/blueprint/components/hashes/sha256/plonk/sha256.hpp>
+#include <nil/blueprint/components/hashes/sha2/plonk/sha256.hpp>
 
 namespace nil {
     namespace blueprint {
