+++ test/hashes/plonk/sha512_process.cpp
@@ -36,7 +36,7 @@
 
 #include <nil/blueprint/blueprint/plonk/circuit.hpp>
 #include <nil/blueprint/blueprint/plonk/assignment.hpp>
-#include <nil/blueprint/components/hashes/sha256/plonk/sha512_process.hpp>
+#include <nil/blueprint/components/hashes/sha2/plonk/sha512_process.hpp>
 
 #include "../../test_plonk_component.hpp"
